revert move of repeat, tolist
